feat: ipad 반응형 디자인
