+++ src/components/Card.jsx
@@ -23,11 +23,14 @@ const Card = ({ pin = false, data = cardData }) => {
       <div
         className={classnames(
           'grid grid-cols-[80px,auto] gap-x-[30px] relative z-10 h-full',
-          'xl:grid-cols-[140px,auto]',
+          'md:grid-cols-[140px,auto]',
         )}
       >
         <div>
-          <a href="#" className="block relative w-full aspect-1 shadow-[20px_20px_40px_rgba(0,0,0,0.1)]">
+          <a
+            href="#"
+            className="block relative w-full aspect-1 shadow-[20px_20px_40px_rgba(0,0,0,0.1)]"
+          >
             <ModifyThumbnail />
             {pin && (
               <div className="absolute flex items-center justify-center text-[19.2px] top-0 left-0 -translate-x-1/2 -translate-y-1/2 w-[40px] h-[40px] rounded-full pin bg-gradient-to-r from-[#7833DC] to-[#CF52E8]">
@@ -40,17 +43,14 @@ const Card = ({ pin = false, data = cardData }) => {
           <div className="w-full">
             <a
               href="#"
-              className={classnames(
-                'text-[16px] break-all',
-                'xl:text-[24px] ',
-              )}
+              className={classnames('text-[16px] break-all', 'md:text-[24px] ')}
             >
               {title}
             </a>
             <div
               className={classnames(
                 'text-[12px] text-[#6f737a] mt-[5px] line-clamp-2 font-normal break-all',
-                'xl:mt-[12px] xl:text-[16px] xl:line-clamp-3 xl:text-[#54575d] min-h-[3.8em]',
+                'md:mt-[12px] md:text-[16px] md:line-clamp-3 md:text-[#54575d] min-h-[3.8em]',
               )}
             >
               {content()}
@@ -59,7 +59,7 @@ const Card = ({ pin = false, data = cardData }) => {
           <div
             className={classnames(
               'text-[10px] text-[#9ba0a8] mt-[5px]',
-              'xl:text-[14px] xl:mt-[18px]',
+              'md:text-[14px] md:mt-[18px]',
             )}
           >
             2022년 4월 12일 오후 6시 53분
@@ -69,7 +69,7 @@ const Card = ({ pin = false, data = cardData }) => {
       <div
         className={classnames(
           'absolute w-full h-[calc(100%+40px)] bg-[#f4f6fa] top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 opacity-0 group-hover:opacity-100 transition-opacity',
-          'xl:rounded-[10px] xl:h-[calc(100%+60px)]',
+          'md:rounded-[10px] md:h-[calc(100%+60px)]',
         )}
       />
     </div>
